gestionBrazoRobotico/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
filelock==3.15.4
idna==3.7
Markdown==3.6
platformdirs==4.2.2
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
